Move stack deletion to cleanup script
Currently if a multinode build times out, the heat stack it created is
not deleted as jenkins aborts the job before it reaches the deletion
step.

This commit moves the stack deletion code into cleanup.sh which can be
called by jenkins in a post build action. This allows us to ensure
cleanup runs even if the job times out.

Connects rcbops/u-suk-dev#918